pr resolves
